change: fix malfunctioning upload media

+++ src/controllers/album.controllers.js
@@ -60,18 +60,20 @@ async function getAlbumById(req, res) {
 
 async function deleteAlbum(req, res) {
 	const { albumId } = req.params
-	const { imagePublicId } = req.body
 	if (!albumId) {
 		return res.status(404).send({ status: 404 })
 	}
 	try {
+		const findGenre = await db.Album.findOne({ _id: albumId }).lean().exec()
+		if (!findGenre) {
+			return res.status(404).send({ status: 404 })
+		}
+		if (findGenre.imagePublicId) await removeMedia(findGenre.imagePublicId, 'image')
 		await deleteCascadeArray(albumId, db.Genre, 'albums')
 		await deleteCascadeArray(albumId, db.Artist, 'albums')
 		await deleteCascadeObject(albumId, db.Track, 'album')
 		await deleteCascadeLikedByUser(albumId, db.User, 'albums')
-		if (imagePublicId) await removeMedia(imagePublicId, 'image')
 		const albumDelete = await db.Album.findOneAndDelete({ _id: albumId }).lean()
-
 		if (!albumDelete) {
 			return res.status(400).send({ status: 400 })
 		}
@@ -97,11 +99,19 @@ async function putAlbumImage(req, res) {
 		return res.status(404).send({ status: 404 })
 	}
 	try {
+		const findAlbum = await db.Album.findOne({ _id: albumId }).lean().exec()
+		if (!findAlbum) {
+			return res.status(404).send({ status: 404 })
+		}
+		if (findAlbum.imagePublicId) await removeMedia(findAlbum.imagePublicId, 'image')
 		const imageUploaded = await uploadImage(req.files.image.tempFilePath, `${cloudinaryConfig.folder}/albumImage`, 250, 250)
+		if (!imageUploaded) {
+			return res.status(402).send({ status: 402 })
+		}
 		const albumStored = await db.Album.findOneAndUpdate(
 			{ _id: albumId },
 			{
-				imageUrl: imageUploaded.url,
+				imageUrl: imageUploaded.secure_url,
 				imagePublicId: imageUploaded.public_id
 			},
 			{ returnOriginal: false }
@@ -109,18 +119,18 @@ async function putAlbumImage(req, res) {
 		if (!albumStored) {
 			return res.status(400).send({ status: 400 })
 		}
-		await fs.unlink(req.files.image.tempFilePath)
 		return res.status(200).send({ status: 200, album: albumStored })
 	} catch (err) {
 		return res.status(500).send({ status: 500, error: err })
+	} finally {
+		await fs.unlink(req.files.image.tempFilePath)
 	}
 }
 
 async function updateAlbum(req, res) {
 	const { albumId } = req.params
-	const { name, genres, artists, tracks, imagePublicId } = req.body
+	const { name, genres, artists, tracks } = req.body
 	try {
-		if (imagePublicId) await removeMedia(imagePublicId, 'image')
 		const albumToUpdate = await db.Album.findByIdAndUpdate({ _id: albumId }, {
 			name, genres, artists, tracks
 		}).lean().exec()
@@ -130,12 +140,13 @@ async function updateAlbum(req, res) {
 		await tracks?.forEach(async track => {
 			await db.Track.findOneAndUpdate(
 				{ _id: track },
-				{ album: albumId }
+				{ album: albumId },
+				{ returnOriginal: false }
 			).lean().exec()
 		});
 		genres && await migrateCascadeArray(genres, db.Genre, 'albums', albumToUpdate._id)
 		artists && await migrateCascadeArray(artists, db.Artist, 'albums', albumToUpdate._id)
-		return res.status(200).send({ status: 200 })
+		return res.status(200).send({ status: 200, album: albumToUpdate })
 	} catch (err) {
 		return res.status(500).send({ status: 500 })
 	}
